# contracts/openapi/careos-foundation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROOTOPATHY CareOS Foundation API",
    "version": "0.21.0",
    "description": "Checked contract for the currently implemented public foundation, browser identity with approved role-level MFA enforcement, approved-registry governed invitations, maker-checker MFA and membership administration, membership-backed organization selection and access projection, and bounded organization-core administration endpoints with the exact approved readiness, organization-profile, registration-identifier, effective address/contact, and versioned international-settings contracts. Unsafe browser operations require an approved Origin or Referer plus the double-submit CSRF cookie/header pair."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "listSelectableOrganizations",
        "description": "Returns only organizations for which the authenticated actor has a currently active and effective membership. A stored selection is a UX preference and is not authorization evidence.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Selectable organizations for the current actor",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationAccess"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/organizations/{organizationId}/governance-responsibilities": {
      "get": {
        "operationId": "getOrganizationGovernanceDirectory",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }],
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Minimum-necessary governance directory", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGovernanceDirectory" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createOrganizationGovernanceResponsibility",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "$ref": "#/components/parameters/BrowserOrigin" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GovernanceResponsibilityWriteRequest" } } } },
        "responses": { "201": { "description": "Responsibility created", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGovernanceDirectory" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "500": { "$ref": "#/components/responses/InternalError" } }
      }
    },
    "/api/v1/organizations/{organizationId}/governance-responsibilities/{responsibilityId}/supersessions": {
      "post": {
        "operationId": "supersedeOrganizationGovernanceResponsibility",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "$ref": "#/components/parameters/ResponsibilityId" }, { "$ref": "#/components/parameters/BrowserOrigin" }, { "$ref": "#/components/parameters/IfMatch" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GovernanceResponsibilityWriteRequest" } } } },
        "responses": { "200": { "description": "Responsibility superseded", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGovernanceDirectory" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "412": { "$ref": "#/components/responses/PreconditionFailed" }, "428": { "$ref": "#/components/responses/PreconditionRequired" }, "500": { "$ref": "#/components/responses/InternalError" } }
      }
    },
    "/api/v1/organizations/{organizationId}/governance-responsibilities/{responsibilityId}/endings": {
      "post": {
        "operationId": "endOrganizationGovernanceResponsibility",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "$ref": "#/components/parameters/ResponsibilityId" }, { "$ref": "#/components/parameters/BrowserOrigin" }, { "$ref": "#/components/parameters/IfMatch" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GovernanceResponsibilityEndRequest" } } } },
        "responses": { "200": { "description": "Responsibility ended", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGovernanceDirectory" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "412": { "$ref": "#/components/responses/PreconditionFailed" }, "428": { "$ref": "#/components/responses/PreconditionRequired" }, "500": { "$ref": "#/components/responses/InternalError" } }
      }
    },
...
    "/api/v1/organizations/{organizationId}/contacts": {
      "get": {
        "operationId": "listOrganizationContacts",
        "description": "Returns governed effective addresses, masked contacts, migration-owned metadata, and live server-authorized actions.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current governed organization address and masked-contact projection",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationContactCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "createOrganizationContact",
        "description": "Creates a governed contact while returning only its masked projection and keeping confidential values out of evidence.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationContactWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization contact created or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationContact"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CAREOS_SESSION",
        "description": "Opaque HttpOnly SameSite=Strict server-side session identifier. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "XSRF-TOKEN",
        "description": "Readable SameSite=Strict double-submit token cookie. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "Must match the raw XSRF-TOKEN cookie value."
      }
    },
    "parameters": {
      "BrowserOrigin": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "description": "Must exactly match a configured CareOS browser origin. A same-origin Referer is accepted when Origin is unavailable.",
        "schema": {
          "type": "string",
          "format": "uri"
        }
      },
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "description": "Organization boundary for every protected business-resource route.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "ResponsibilityId": {
        "name": "responsibilityId",
        "in": "path",
        "required": true,
        "description": "Opaque governance-responsibility UUID.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Strong entity tag from the latest representation. Required for protected updates and deletes.",
        "schema": {
          "type": "string",
          "pattern": "^\"[A-Za-z0-9._:-]{1,128}\"$"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller-generated key for an explicitly retryable protected mutation. Reuse is valid only for an equivalent request.",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      }
    },
    "headers": {
      "CorrelationId": {
        "description": "Stable identifier used to correlate one HTTP request across logs and problem responses.",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
...
      "ETag": {
        "description": "Strong entity tag for optimistic concurrency.",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\"[A-Za-z0-9._:-]{1,128}\"$"
        }
      },
      "SessionExpiresIn": {
        "description": "Effective seconds remaining before the authenticated browser session reaches its idle or absolute deadline. Omitted when no authenticated session exists.",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 2147483647
        }
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Request validation, password policy, or one-time-token failure",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Credentials, verification evidence, or session is invalid",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Origin, CSRF, or authorization check failed",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource is unavailable or hidden from the current actor",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "The requested transition, idempotency key, or current resource state conflicts with the operation",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required precondition is missing, such as recent authentication, recent MFA, or If-Match",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied entity tag no longer matches the current representation",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request failed without exposing sensitive implementation details.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "GovernanceAssignee": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "type", "display"],
        "properties": { "id": { "type": "string", "format": "uuid" }, "type": { "type": "string", "enum": ["membership", "external_contact"] }, "display": { "type": "string", "minLength": 1, "maxLength": 200 } }
      },
...
      "OrganizationGovernanceDirectory": {
        "type": "object", "additionalProperties": false,
        "required": ["organizationId", "canManage", "evaluatedAt", "responsibilityTypes", "eligibleAssignees", "responsibilities"],
        "properties": {
          "organizationId": { "type": "string", "format": "uuid" }, "canManage": { "type": "boolean" }, "evaluatedAt": { "type": "string", "format": "date-time" },
          "responsibilityTypes": { "type": "array", "prefixItems": [{ "const": "clinical" }, { "const": "privacy" }, { "const": "security" }, { "const": "billing" }], "minItems": 4, "maxItems": 4 },
          "eligibleAssignees": { "type": "array", "items": { "$ref": "#/components/schemas/GovernanceAssignee" } },
          "responsibilities": { "type": "array", "items": { "$ref": "#/components/schemas/GovernanceResponsibility" } }
        }
      },
      "GovernanceResponsibilityWriteRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["responsibilityType", "membershipId", "externalContactId", "escalationEmail", "escalationPhone", "effectiveFrom", "reason"],
        "properties": {
          "responsibilityType": { "type": "string", "enum": ["clinical", "privacy", "security", "billing"] },
          "membershipId": { "type": ["string", "null"], "format": "uuid" }, "externalContactId": { "type": ["string", "null"], "format": "uuid" },
          "escalationEmail": { "type": ["string", "null"], "maxLength": 254 }, "escalationPhone": { "type": ["string", "null"], "maxLength": 32 },
          "effectiveFrom": { "type": "string", "format": "date-time" }, "reason": { "type": "string", "minLength": 10, "maxLength": 500 }
        }
      },
      "GovernanceResponsibilityEndRequest": {
        "type": "object", "additionalProperties": false, "required": ["effectiveTo", "reason"],
        "properties": { "effectiveTo": { "type": "string", "format": "date-time" }, "reason": { "type": "string", "minLength": 10, "maxLength": 500 } }
      },
...
      "OrganizationContact": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "contactId",
          "channel",
          "purpose",
          "purposeDisplayName",
          "maskedValue",
          "verificationStatus",
          "isPrimary",
          "isPreferred",
          "effectiveFrom",
          "effectiveTo",
          "supersedesId",
          "status",
          "availableActions",
          "lockVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "contactId": {
            "type": "string",
            "format": "uuid"
          },
          "channel": {
            "type": "string",
            "enum": ["email", "phone", "web"]
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z][a-z0-9]*([._:-][a-z0-9]+)*$"
          },
          "purposeDisplayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "maskedValue": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048,
            "description": "Minimum-necessary masked projection; the confidential normalized value is never returned."
          },
          "verificationStatus": {
            "type": "string",
            "enum": ["unverified", "verified"]
          },
          "isPrimary": {
            "type": "boolean"
          },
          "isPreferred": {
            "type": "boolean"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          },
          "effectiveTo": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "supersedesId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["scheduled", "active", "ended", "superseded"]
          },
          "availableActions": {
            "type": "array",
            "maxItems": 3,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": ["verify", "supersede", "end"]
            }
          },
          "lockVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrganizationContactCollection": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationId",
          "canCreate",
          "addressTypes",
          "purposes",
          "addresses",
          "contacts"
        ],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "canCreate": {
            "type": "boolean"
          },
          "addressTypes": {
            "type": "array",
            "minItems": 4,
            "maxItems": 4,
            "uniqueItems": true,
            "prefixItems": [
              { "const": "registered" },
              { "const": "postal" },
              { "const": "service" },
              { "const": "billing" }
            ],
            "items": false
          },
          "purposes": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/OrganizationContactPurpose"
            }
          },
          "addresses": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/OrganizationAddress"
            }
          },
          "contacts": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/OrganizationContact"
            }
          }
        }
      },
...
      "OrganizationContactWriteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "channel",
          "purpose",
          "value",
          "isPrimary",
          "isPreferred",
          "effectiveFrom",
          "effectiveTo",
          "reason"
        ],
        "properties": {
          "channel": {
            "type": "string",
            "enum": ["email", "phone", "web"]
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z][a-z0-9]*([._:-][a-z0-9]+)*$"
          },
          "value": {
            "type": "string",
            "minLength": 3,
            "maxLength": 2048,
            "description": "Confidential value normalized and validated according to channel; never projected back."
          },
          "isPrimary": {
            "type": "boolean"
          },
          "isPreferred": {
            "type": "boolean"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          },
          "effectiveTo": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          }
        }
      },
...
      "OrganizationAccess": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "status", "roleKeys", "selected"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active"]
          },
          "roleKeys": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "selected": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}